+++ pyagentic/_base/_exceptions.py
